embedder/requirements.txt
psycopg2-binary
python-dotenv
pandas
sqlalchemy
sentence-transformers